# Reuse source buffers