--- a/testing/test-cases/selenium-tests/glLinesSpeed/include.html
+++ b/testing/test-cases/selenium-tests/glLinesSpeed/include.html
@@ -1,0 +1,1 @@
+<div id='map'></div>