--- a/testing/test-cases/selenium-tests/glPointsSpeed/include.css
+++ b/testing/test-cases/selenium-tests/glPointsSpeed/include.css
@@ -1,0 +1,3 @@
+body {
+    overflow: hidden;
+}