This is a further extension of the `allocate-less-memory` branch, but the changes introduced in the last commit are independent of that, so I'm adding this as a separate PR.  If this is reviewed directly, both can be satisfied with this PR.

Rather than using the DataBuffers structure, allocate source buffers directly, reusing them if they haven't changed.  Also, for points and lines, we require the vertex indices to exist, but we don't do anything with them, so create but don't fill those arrays.

This shouldn't speed up the first load of data much, but when data is reloaded with the same number of points or lines, it improves the speed.  Compared to allocating the buffers every time, this reduces the time to load 1,000,000 points from 0.8 s to 0.7 s and for 1,000,000 lines from 4.0 s to 3.5 s (roughly 12 % improvement).