@jbeezley 2015-03-11
The test look great.  I'll leave the review of the GL to @aashish24.